serve/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
requests==2.32.3
httpx==0.27.2
beautifulsoup4==4.12.3
trafilatura==1.12.2
pydantic==2.9.2
PyYAML==6.0.2
openai==1.51.2
jinja2==3.1.4
python-dotenv==1.0.1
azure-storage-blob==12.23.1
